Sample .sbatch file for an array job
This type of file can be used to launch multiple jobs simultaneously. When you run the .sbatch script, each job is launched and is assigned a  TASK_ID. I typically make a table in R with rows corresponding to the TASK_ID's and columns to the parameters that define each run (e.g. forecast date, training period, iterations to run, etc).  I then save this table as a .rda file and add code to my main .R script that reads in the .rda file and assigns the parameters from the row corresponding to the TASK_ID as variable. This workflow is nice, because if you change what parameters or how many tasks you want to consider, you only need to update the .rda and the <last_TASK_ID>.